Conflicts:
	aplicacao/contest/src/main/java/ufc/quixada/npi/contest/repository/PessoaRepository.java
	aplicacao/contest/src/main/java/ufc/quixada/npi/contest/service/PessoaService.java

+++ aplicacao/contest/src/test/java/ufc/quixada/npi/contest/AlterarEventoInativoRequestSteps.java
@@ -1,18 +1,12 @@
 package ufc.quixada.npi.contest;
 
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+import static org.mockito.Mockito.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
@@ -21,7 +15,6 @@ import cucumber.api.java.pt.*;
 import ufc.quixada.npi.contest.controller.EventoController;
 import ufc.quixada.npi.contest.model.EstadoEvento;
 import ufc.quixada.npi.contest.model.Evento;
-import ufc.quixada.npi.contest.model.PapelEvento;
 import ufc.quixada.npi.contest.model.Pessoa;
 import ufc.quixada.npi.contest.service.EventoService;
 import ufc.quixada.npi.contest.service.ParticipacaoEventoService;
@@ -69,7 +62,7 @@ public class AlterarEventoInativoRequestSteps {
 	public void administradorAlteraEventoInexistente(String nome) throws Throwable{
 		evento = new Evento();
 		evento.setNome(nome);
-		evento.setId(Integer.valueOf(EVENTO_ID));
+		evento.setId(Long.valueOf(EVENTO_ID));
 		evento.setEstado(EstadoEvento.INATIVO);
 		
 		when(eventoService.existeEvento(evento.getId())).thenReturn(true);
